backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1

# File validation
python-magic==0.4.27
aiofiles==23.2.1

# NLP and AI
spacy==3.8.2
openai==1.54.0  # OpenRouter AI compatible client
# anthropic==0.39.0  # Not currently used

# PDF Generation
weasyprint==60.1
pydyf<0.11.0  # Pin to <0.11.0 for compatibility with weasyprint 60.1
Jinja2==3.1.2

# Security and Rate Limiting
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
markdown==3.5.1

# CORS (built into FastAPI, no additional package needed)
